-moved components to their own files
