+++ lexicon/src/components/index.ts
@@ -1,4 +1,7 @@
 export * from './AddWord';
+export * from './Definition';
+export * from './DefinitionElement';
 export * from './SearchBar';
 export * from './Tab';
 export * from './Vote';
+export * from './WordItem';
